Fix backend cookie and CORS config, move to config/ folder, update frontend auth with lib/auth.ts and context/AuthContext.tsx, add lib/types.ts, ensure localhost compatibility
