+++ backend/auth-service/api/config/cors.config.ts
@@ -1,0 +1,10 @@
+import { CorsOptions } from "cors";
+
+const corsOptions: CorsOptions = {
+  origin: process.env.CLIENT_URL || "http://localhost:3000",
+  credentials: true,
+  methods: ["GET", "POST", "PUT", "DELETE", "OPTIONS"],
+  allowedHeaders: ["Content-Type", "Authorization"],
+};
+
+export default corsOptions;
