+++ frontend/package.json
@@ -40,8 +40,9 @@
     "@radix-ui/react-tooltip": "^1.1.6",
     "@tsparticles/engine": "^3.8.1",
     "@tsparticles/react": "^3.0.0",
+    "@types/js-cookie": "^3.0.6",
     "autoprefixer": "^10.4.20",
-    "axios": "^1.8.4",
+    "axios": "^1.9.0",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
     "cmdk": "1.0.4",
@@ -49,6 +50,7 @@
     "embla-carousel-react": "8.5.1",
     "framer-motion": "^12.7.4",
     "input-otp": "1.4.1",
+    "js-cookie": "^3.0.5",
     "jsonwebtoken": "^9.0.2",
     "lucide-react": "^0.454.0",
     "next": "15.2.4",
